Add preliminary downloader
The downloader downloads files from localhost/nosemod/. Any files in there will be downloaded. The downloader currently relies on Apache directory listings. However, any script, say PHP, that outputs <a> tags with the names of the files could work, so it's fairly portable. The only difference for the Apache native dirlisting is that the code has to ignore its links (like "Name", "Parent Directory", etc).